Complete implementation of the block manager. It loads and unload data from disk according to a priority that can be changed on run-time
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@1235 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ libs/engine/DiskOperation.cpp
@@ -349,9 +349,10 @@ namespace engine {
 				setError("Error while removing file");
 		}
         
-        
+
 		// If write operation, destroy the buffer ( only in write and append operations )
-        destroyBuffer();
+        if( environment.get( destroy_buffer_after_write , "yes") == "yes" )
+            destroyBuffer();
 		
 		// Notify to the engine
 		diskManager->finishDiskOperation( this );
